LayoutTests/platform/ios-simulator/fast/events/touch/gesture/gesture-dblclick-expected.txt

CONSOLE MESSAGE: line 75: TypeError: undefined is not a function (evaluating 'eventSender.gestureTap(10, 12)')
FAIL: Timed out waiting for notifyDone to be called
This tests gesture callbacks for a double tap sequence of gestures.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
